Iniciando jogo

+++ funcionario.h
@@ -16,15 +16,15 @@ using namespace std;
 class Funcionario() {
 	private:
 		string m_nome;
-		float m_salario;
+		double m_salario;
 		Data m_admissao; // Instancia um objeto da classe Data.
 	public:
 		string getNome();
-		float getSalario();
+		double getSalario();
 		Data getAdmissao();
 
 		Funcionario();
-		Funcionario(string nome_, float salario_, Data admissao_);
+		Funcionario(string nome_, double salario_, Data admissao_);
 		~Funcionario();
 
 		void statusSalario(float porcentagem_);
